consumer/requirements.txt
aiokafka
openai
pinecone-client
tenacity
python-dotenv
